Support `group_by()` with `count()` for GBIF (#272)
First pass at this, as it still only supports one facet value. Note that `show_values()` for type `fields` could benefit from using the occurrences/search API, which will need to be recoded. That API also supports longer facet results and so may be useful for building queries with >1 facet.